--- adoggy-app/src/components/DogCards.css
@@ -1,66 +1,0 @@
-.card {
-    border-radius: 60px;
-    position: relative;
-    background-color: #fff;
-    width: 600px;
-    padding: 20px;
-    max-width: 85vw;
-    height: 60vh;
-    box-shadow: 0px 18px 53px 0px rgba(0, 0, 0, 0.3);
-    background-size: cover;
-    background-position: center;
-    border-top-right-radius: 0;
-    border-bottom-left-radius: 0;
-    border: 1px solid black;
-}
-
-.dog-cards-container {
-    display: flex;
-    justify-content: center;
-    margin-top: 5vh;
-}
-
-
-.swipe {
-    position: absolute;
-}
-
-.card h3 {
-    position: absolute;
-    bottom: 0;
-    margin: 10px;
-    color: #fff;
-    font-size: 25px;
-    font-weight: bold;
-    -webkit-text-stroke: 1px black;
-}
-
-.card h3:nth-child(1) {
-    left: 0;
-}
-
-.card h3:nth-child(2) {
-    position: absolute;
-    bottom: 0;
-    right: 20px;
-    color: #fff;
-    font-size: 25px;
-    font-weight: bold;
-    -webkit-text-stroke: 1px black;
-}
-
-.card h3:nth-child(3) {
-    position: absolute;
-    top: -10px;
-    right: -5px;
-    color: #fff;
-    font-size: 25px;
-    font-weight: bold;
-    -webkit-text-stroke: 1px black;
-}
-
-.dog-cards {
-    height: 100%;
-    width: 100%;
-}
-
