--- adoggy-app/src/components/Header.css
@@ -1,8 +1,0 @@
-
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    z-index: 100;
-    border-bottom: 1px solid lightgray;
-}
