--- adoggy-app/src/components/SwipeButtons.css
@@ -1,23 +1,0 @@
-.swipe-buttons {
-    position: fixed;
-    bottom: 2vh;
-    display: flex;
-    width: 100%;
-    padding-top: 20px;
-    justify-content: space-evenly;
-}
-
-
-.swipe-buttons .MuiIconButton-root {
-    background-color: white;
-    box-shadow: 0px 10px 53px 0px rgba(0, 0, 0, 0.3) !important;
-}
-
-.thumbs-up {
-    
-    color: rgb(38, 64, 179);
-}
-
-.thumbs-down {
-    color: rgb(160, 40, 40)
-}
