--- adoggy-app/src/components/axios.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-const instance = axios.create({
-    baseUrl: 'https://adopts-backend.herokuapp.com/'
-})
-
-export default instance;
