+++ package.json
@@ -11,6 +11,8 @@
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-tinder-card": "^1.4.0",
     "web-vitals": "^1.1.2"
