+++ src/App.js
@@ -1,17 +1,24 @@
-import React from 'react';
+import React, {  useEffect, createContext, useState } from 'react';
 import Header from './components/Header';
 import DogCards from './components/DogCards';
 import SwipeButtons from './components/SwipeButtons';
 import './App.css';
 
-function App() {
-  return (
-    <div className="app">
 
+
+export const AuthContext = createContext();
+
+function App() {
+
+  return (
+    <>
+      
       <Header />
       <DogCards />
       <SwipeButtons />
-    </div>
+    
+   </>
+    
   );
 }
