+++ src/components/DogCards.css
@@ -4,9 +4,9 @@
     background-color: #fff;
     width: 600px;
     padding: 20px;
-    max-width: 85vw;
+    max-width: 80vw;
     height: 60vh;
-    box-shadow: 0px 18px 53px 0px rgba(0, 0, 0, 0.3);
+    box-shadow: 0px 10px 15px 0px rgba(75, 75, 75, 0.356);
     background-size: cover;
     background-position: center;
     border-top-right-radius: 0;
